new try

+++ exercises.c
@@ -81,7 +81,7 @@ posiciona en el elemento anterior.
 
 void eliminaElementos(List*L, int elem)
 {
-  int *elemento = first(L);
+  float *elemento = first(L);
   while(elemento != NULL)
     {
       if (*(int*)elemento == elem)
